Added Conditional for Deleting Keys at end of program
Only Deletes keys if they exist.

+++ MainApp.java
@@ -104,7 +104,9 @@ public class MainApp extends Application{
 	}
 
 	private void closeProgram(){
+		if(RSA.doesExist()){
 		RSA.deleteKeys();
+		}
 		System.gc();
 		try {
 			if(Database.connected == true){
